inventory-microservice/pyproject.toml
[tool.poetry]
name = "micro-service-app"
version = "0.1.0"
description = ""
authors = ["askat <askat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
redis-om = "^0.3.2"
uvicorn = "^0.31.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

